[BE:fix] out of boundary시 get arrest가 여러번 나가는 것을 수정

See merge request s14-webmobile4-sub1/S14P11D104!227

+++ be/pnt_node/websocket_server/src/socket/games/controller/GameController.js
@@ -294,7 +294,7 @@ export class GameController {
             locationData.penalty = res.penalty;
             await this.redisClient.setLocation(memberId, gameId, locationData);
 
-            if (count >= 3) {
+            if (count >= 3 && status === GameMemberStatus.FREE) {
                 locationData.status = GameMemberStatus.TRANSFER;
                 await this.redisClient.deletePenalty(memberId, gameId);
